Move CLI utility code to the core library
